+++ app/src/main/AndroidManifest.xml
@@ -141,7 +141,7 @@
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar"/>
 
-        <activity android:name=".src.main.seller.store.productfeed.SellerProductDetailFeedActivity"
+        <activity android:name=".src.main.seller.store.productfeed.productdetail.SellerProductDetailFeedActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar"/>
@@ -182,7 +182,7 @@
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
 
-        <activity android:name=".src.main.seller.store.productfeed.product.SellerProductAddActivity"
+        <activity android:name=".src.main.seller.store.productfeed.productadd.SellerProductAddActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
@@ -194,7 +194,7 @@
             android:theme="@style/AppTheme2">
         </activity>
 
-        <activity android:name=".src.main.seller.store.productfeed.product.SellerProductEditActivity"
+        <activity android:name=".src.main.seller.store.productfeed.productadd.SellerProductEditActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme2">
