--- app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/model/Image.kt
@@ -1,6 +1,0 @@
-package com.codepatissier.keki.src.main.seller.store.productfeed.model
-
-data class Image(
-    val postIdx : Long,
-    val postImgUrl : String
-)
