+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/productadd/SellerProductAddService.kt
@@ -1,8 +1,8 @@
-package com.codepatissier.keki.src.main.seller.store.productfeed.product
+package com.codepatissier.keki.src.main.seller.store.productfeed.productadd
 
 import com.codepatissier.keki.config.ApplicationClass
-import com.codepatissier.keki.src.main.seller.store.productfeed.product.model.SellerProductAddBody
-import com.codepatissier.keki.src.main.seller.store.productfeed.product.model.SellerProductAddResponse
+import com.codepatissier.keki.src.main.seller.store.productfeed.productadd.model.SellerProductAddBody
+import com.codepatissier.keki.src.main.seller.store.productfeed.productadd.model.SellerProductAddResponse
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
