+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/productadd/SellerProductAddView.kt
@@ -1,0 +1,8 @@
+package com.codepatissier.keki.src.main.seller.store.productfeed.productadd
+
+import com.codepatissier.keki.src.main.seller.store.productfeed.productadd.model.SellerProductAddResponse
+
+interface SellerProductAddView {
+    fun onPostProductSuccess(response: SellerProductAddResponse)
+    fun onPostProductFailure(message:String)
+}
