+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/productadd/model/SellerProductAddResponse.kt
@@ -1,4 +1,4 @@
-package com.codepatissier.keki.src.main.seller.store.productfeed.product.model
+package com.codepatissier.keki.src.main.seller.store.productfeed.productadd.model
 
 data class SellerProductAddResponse(
     val code: Int,
