+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/productdetail/DetailImageAdapter.kt
@@ -1,9 +1,8 @@
-package com.codepatissier.keki.src.main.seller.store.productfeed
+package com.codepatissier.keki.src.main.seller.store.productfeed.productdetail
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.google.firebase.storage.FirebaseStorage
 
 class DetailImageAdapter(fa: FragmentActivity, val images: Array<String?>): FragmentStateAdapter(fa) {
