+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/productdetail/SellerProductFeedDetailRetrofitInterface.kt
@@ -1,10 +1,8 @@
-package com.codepatissier.keki.src.main.seller.store.productfeed
+package com.codepatissier.keki.src.main.seller.store.productfeed.productdetail
 
 import com.codepatissier.keki.config.BaseResponse
-import com.codepatissier.keki.src.main.consumer.store.productfeed.productdetail.model.ConsumerStoreProductFeedDetailResponse
-import com.codepatissier.keki.src.main.seller.store.productfeed.product.model.SellerProductAddBody
-import com.codepatissier.keki.src.main.seller.store.productfeed.product.model.SellerProductAddResponse
-import com.codepatissier.keki.src.main.seller.store.productfeed.model.SellerProductFeedDetailResponse
+import com.codepatissier.keki.src.main.seller.store.productfeed.productadd.model.SellerProductAddBody
+import com.codepatissier.keki.src.main.seller.store.productfeed.productdetail.model.SellerProductFeedDetailResponse
 import retrofit2.Call
 import retrofit2.http.*
