+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/productdetail/SellerProductFeedDetailService.kt
@@ -1,13 +1,12 @@
-package com.codepatissier.keki.src.main.seller.store.productfeed
+package com.codepatissier.keki.src.main.seller.store.productfeed.productdetail
 
 import com.codepatissier.keki.config.ApplicationClass
 import com.codepatissier.keki.config.BaseResponse
-import com.codepatissier.keki.src.main.seller.store.productfeed.product.model.SellerProductAddBody
-import com.codepatissier.keki.src.main.seller.store.productfeed.model.SellerProductFeedDetailResponse
+import com.codepatissier.keki.src.main.seller.store.productfeed.productadd.model.SellerProductAddBody
+import com.codepatissier.keki.src.main.seller.store.productfeed.productdetail.model.SellerProductFeedDetailResponse
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-import retrofit2.http.Body
 
 class SellerProductFeedDetailService(val sellerProductFeedDetailView: SellerProductFeedDetailView) {
     fun tryGetProductFeedDetail(dessertIdx:Long){
