+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/productdetail/model/Image.kt
@@ -1,0 +1,6 @@
+package com.codepatissier.keki.src.main.seller.store.productfeed.productdetail.model
+
+data class Image(
+    val postIdx : Long,
+    val postImgUrl : String
+)
