+++ app/src/main/java/com/codepatissier/keki/src/main/seller/store/productfeed/productdetail/model/SellerProductFeedDetailResponse.kt
@@ -1,4 +1,4 @@
-package com.codepatissier.keki.src.main.seller.store.productfeed.model
+package com.codepatissier.keki.src.main.seller.store.productfeed.productdetail.model
 
 data class SellerProductFeedDetailResponse(
     val isSuccess: Boolean,
